requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
